+++ extensions/deferral/rvr/src/lib.rs
@@ -160,7 +160,7 @@ impl DeferralRvrExtension {
 }
 
 impl<F: PrimeField32> RvrExtension<F> for DeferralRvrExtension {
-    fn try_lift(&self, insn: &Instruction<F>, pc: u32) -> Option<LiftedInstr> {
+    fn try_lift(&self, insn: &Instruction<F>, pc: u64) -> Option<LiftedInstr> {
         let opcode = insn.opcode.as_usize();
 
         if opcode == DeferralOpcode::CALL.global_opcode_usize() {
